+++ source/scene/CylinderShape.h
@@ -17,7 +17,7 @@ namespace scene
         ~CCylinderShape();
 
         void        render()         override;
-        void        update(f64 time) override;
+        void        update(s32 time) override;
 
         void        init()           override;
